+++ Examples/CharacterControllerMovingBodies/Assets/Scripts/RotatingBodyLogic.cs
@@ -3,7 +3,7 @@ using Unity.Netcode;
 using Unity.Netcode.Components;
 using UnityEngine;
 #if UNITY_EDITOR
-using Unity.Netcode.Editor;
+using Unity.Netcode.GameObjects.Editor;
 using UnityEditor;
 
 /// <summary>