+++ com.unity.netcode.gameobjects/Tests/Editor/Timing/TimingTestHelper.cs
@@ -60,4 +60,3 @@ namespace Unity.Netcode.GameObjects.EditorTests
         }
     }
 }
-